PYTHON: Move NumPy specific tests to test_array_numpy.py
